# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: API Gestionnaire de Bourse ALPE
  version: 1.0.0
  description: |
    API REST pour l'application de gestion des bourses aux vêtements et jouets
    organisées par l'association ALPE Plaisance du Touch.

    ## Authentification
    L'API utilise des tokens JWT (Bearer). Après connexion, incluez le token
    dans l'en-tête `Authorization: Bearer <token>`.

    ## Rôles
    - **deposant** : Gère ses propres listes et articles
    - **benevole** : Scanne et encaisse les ventes
    - **gestionnaire** : Configure les éditions, génère les étiquettes
    - **admin** : Gestion complète (utilisateurs, éditions)

    ## Codes d'erreur
    - 400 : Requête invalide (validation)
    - 401 : Non authentifié
    - 403 : Accès refusé (permissions)
    - 404 : Ressource non trouvée
    - 409 : Conflit (doublon, état invalide)
    - 422 : Entité non traitable (règle métier violée)
    - 429 : Trop de requêtes (rate limiting)
  contact:
    name: ALPE Plaisance du Touch
    email: contact@alpe-plaisance.fr
  license:
    name: Propriétaire
    url: https://alpe-plaisance.fr
servers:
  - url: https://api.bourse-alpe.fr/v1
    description: Production
  - url: https://staging-api.bourse-alpe.fr/v1
    description: Staging
  - url: http://localhost:8000/v1
    description: Développement local
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Connexion utilisateur
      description: Authentifie un utilisateur et retourne les tokens JWT.
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: marie.dupont@email.fr
                password:
                  type: string
                  format: password
                  minLength: 8
      responses:
        '200':
          description: Connexion réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '401':
          description: Identifiants invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Trop de tentatives (blocage 15 min)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rafraîchir le token d'accès
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Nouveau token généré
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
        '401':
          description: Refresh token invalide ou expiré
...
  /auth/activate:
    post:
      tags: [Auth]
      summary: Activer un compte via invitation
      description: |
        Active un compte déposant à partir d'un token d'invitation.
        Le token est à usage unique et expire après 7 jours.
      operationId: activateAccount
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token:
                  type: string
                  description: Token d'invitation reçu par email
                password:
                  type: string
                  format: password
                  minLength: 8
                  description: "Min 8 car., 1 lettre, 1 chiffre, 1 spécial"
                nom:
                  type: string
                prenom:
                  type: string
                telephone:
                  type: string
      responses:
        '201':
          description: Compte activé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Token invalide ou expiré
        '409':
          description: Compte déjà activé
  /auth/password/reset-request:
    post:
      tags: [Auth]
      summary: Demander réinitialisation mot de passe
      operationId: requestPasswordReset
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        '202':
          description: Email envoyé si le compte existe
  /auth/password/reset:
    post:
      tags: [Auth]
      summary: Réinitialiser le mot de passe
      operationId: resetPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token:
                  type: string
                password:
                  type: string
                  format: password
                  minLength: 8
      responses:
        '200':
          description: Mot de passe réinitialisé
        '400':
          description: Token invalide ou expiré

  # ============================================================================
  # ÉDITIONS
  # ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    UserRole:
      type: string
      enum: [deposant, benevole, gestionnaire, admin]
...
    AuthTokens:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          default: bearer
        expires_in:
          type: integer
          description: Durée de validité en secondes
        user:
          $ref: '#/components/schemas/User'

    # --------------------------------------------------------------------------
    # UTILISATEURS
    # --------------------------------------------------------------------------
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        nom:
          type: string
        prenom:
          type: string
        telephone:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        created_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
security:
  - bearerAuth: []